+++ crates/roux-core/src/models/mod.rs
@@ -6,6 +6,7 @@ mod layout;
 mod notification;
 mod profile;
 mod project;
+mod pty;
 mod session;
 mod settings;
 mod subscription;
@@ -37,6 +38,7 @@ pub use notification::{
 };
 pub use profile::{ProfileSource, Provider, SpawnProfile, StartupBehavior};
 pub use project::{Project, ProjectUpdate, SessionBlueprint};
+pub use pty::{PtyInfo, PtyRole, PtyStatus};
 pub use session::{Session, SessionStatus};
 pub use settings::{
     CursorStyle, ExampleVariant, ExperimentsConfig, GroupBy, LibrarySource, LibrarySourceKind,
