+++ src-tauri/Cargo.toml
@@ -55,6 +55,7 @@ rmcp = { version = "1.6.0", features = ["server", "transport-io", "schemars"] }
 roux-core = { path = "../crates/roux-core" }
 roux-git = { path = "../crates/roux-git" }
 roux-gh = { path = "../crates/roux-gh" }
+roux-runtime = { path = "../crates/roux-runtime" }
 roux-smolvm = { path = "../crates/roux-smolvm" }
 roux-worktrunk = { path = "../crates/roux-worktrunk" }
 schemars = "1"
